+++ js/ui.js
@@ -1985,16 +1985,26 @@
 		var parameters = [];
 		if (instruction.parameters !== null) {
 			var i;
+			// First we search which is the last parameter with value, so we use "undefined" with all unset
+			// parameters up to it and don't set the parameters at all after it
+			var lastParameterWithValue = -1;
+			for (var i=instruction.parameters.length-1; i>=0 && lastParameterWithValue<0; i--) {
+				if (div.getAttribute("param"+(i+1)) != null && div.getAttribute("param"+(i+1)) !== "") {
+					lastParameterWithValue = i;
+				}
+			}
 			for (i=0; i<instruction.parameters.length; i++) {
 				var param = undefined;
 				if (div.getAttribute("param"+(i+1)) !== null) {
 					param = div.getAttribute("param"+(i+1));
-				} else if (instruction.parameters[i].initial !== undefined) {
+				} else if (instruction.parameters[i].initial !== undefined && instruction.parameters[i].optional !== true) {
 					param = instruction.parameters[i].initial;
 					if (instruction.parameters[i].type == "number") {
 						param = $e_parsePredefinedConstants(param);
 					}
 					div.setAttribute("param"+(i+1),param);
+				} else if (i <= lastParameterWithValue) {
+					param = "";
 				}
 				if (param !== undefined) {
 					parameters.push(param);
@@ -2034,6 +2044,7 @@
 			var bracketsStatus = null; // null indicates we haven't started yet
 			var bracketsExist = false;
 			for (i=0; i<parameters.length; i++) {
+				console.log(i)
 				if (!bracketsStatus && !instruction.parameters[i].noBrackets) {
 					if (instruction.code && instruction.code.space && text[text.length-1] !== " ") {
 						text += " ";
@@ -2041,7 +2052,7 @@
 					text += "(";
 					bracketsStatus = true;
 					bracketsExist = true;
-				} else if (bracketsStatus && parameters[i].noBrackets) {
+				} else if (bracketsStatus && instruction.parameters[i].noBrackets) {
 					text += ") ";
 					bracketsStatus = false;
 				} else if (bracketsStatus !== null) {